Create / Rename memory management source files
