--- kernel/mem/MemoryManager.cpp
@@ -1,6 +1,0 @@
-#include <mem/MemoryManager.hpp>
-#include <mem/MemoryRange.hpp>
-#include <lib/string.hpp>
-#include <dev/serial/rs232.hpp>
-
-
